Move IconCollectionTable back to terra-core-docs
