+++ packages/terra-core-docs/src/terra-dev-site/doc/icon/IconCollectionTable/IconCollectionRow.module.scss
