+++ packages/terra-core-docs/src/terra-dev-site/doc/icon/IconCollectionTable/IconCollectionTable.module.scss
@@ -1,4 +1,4 @@
-@import '../clinical-lowlight-theme/IconCollectionTable.module';
+@import './clinical-lowlight-theme/IconCollectionTable.module';
 
 .scroll-container {
   overflow-x: scroll;
