+++ packages/terra-core-docs/src/terra-dev-site/doc/icon/IconCollectionTable/clinical-lowlight-theme/IconCollectionTable.module.scss
