+++ packages/terra-core-docs/src/terra-dev-site/doc/icon/IconCollectionTable/index.js
