--- packages/terra-core-docs/src/terra-dev-site/doc/icon/Themeable.3.doc.mdx
@@ -1,19 +1,0 @@
-import { Badge } from 'terra-icon/package.json?dev-site-package';
-
-import IconThemeable from './example/IconThemeable';
-
-<Badge />
-
-# Terra Icon - Themeable
-
-The following are examples of themeable icons within the terra-icon package.
-
-Note: All icons that were previously exported as 'low light' icons have been deprecated and removed from our examples in favor of using the lowlight theme.
-
-## Component Features
-* [Cross-Browser Support](https://engineering.cerner.com/terra-ui/about/terra-ui/component-standards#cross-browser-support)
-* [Responsive Support](https://engineering.cerner.com/terra-ui/about/terra-ui/component-standards#responsive-support)
-* [Mobile Support](https://engineering.cerner.com/terra-ui/about/terra-ui/component-standards#mobile-support)
-
-## Examples
-<IconThemeable />
