+++ packages/terra-core-docs/tests/jest/iconCollectionData.jsx
