+++ packages/terra-core-docs/tests/wdio/iconCollectionTable-spec.js
