+++ packages/terra-docs/src/index.js
@@ -1,4 +1,3 @@
 export { default as Notice } from './Notice';
 export { default as Placeholder } from './Placeholder';
 export { default as Button } from './Button';
-export { default as IconCollectionTable } from './IconCollectionTable';
